+++ packages/form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tamagui-extras/form
 
+## 1.13.3
+
+### Patch Changes
+
+- fix datepicker native and improve checkbox element
+
 ## 1.13.2
 
 ### Patch Changes